Weather Checker Agent/back-end/pyproject.toml
[project]
name = "back-end"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-google-genai>=2.0.10",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.7",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.12",
    "uvicorn>=0.35.0",
]
